+++ _scripts/discord/cmd/add.js
@@ -27,7 +27,33 @@ module.exports = {
     const checkUserpromise = checkUser(user_info);
     const getBalance = wallet.GetBalance;
 
-    const messageString = JSON.stringify(message);
+
+    // use to send a reply to user with delay and stop typing
+    // ReplyMessage(' Check your DM\'s');
+    function ReplyMessage(content, message) {
+      console.log(JSON.stringify(message));
+      message.channel.startTyping();
+      setTimeout(function() {
+        message.reply(content);
+        message.channel.stopTyping(true);
+      }, 500);
+    }
+
+    // errorMessage({ error: 'Can\'t access faucet from DM!', description: 'Please try again from the main chat, this function will only work there.' });
+    function ErrorMessage(content, footer = '  .: Tipbot provided by The QRL Contributors :.') {
+      console.log(`discord-hellpers: ${JSON.stringify(message)}`);
+      message.channel.startTyping();
+      setTimeout(function() {
+        const embed = new Discord.MessageEmbed()
+          .setColor(0x000000)
+          .setTitle(':warning:  ERROR: ' + content.error)
+          .setDescription(content.description)
+          .setFooter(footer);
+        message.reply({ embed });
+        message.channel.stopTyping(true);
+      }, 1000);
+    }
+
 
     function toQuanta(number) {
       const shor = 1000000000;
@@ -70,7 +96,7 @@ module.exports = {
         if (found === 'true') {
           // user is found, have they been banned?
           if (result.banned) {
-            discordHelpers.errorMessage({ error: 'User is Banned...', description: 'The user has been banned from the tipbot and cannot use the service.\n User Banned on `' + result.banned_date + '`' }, messageString);
+            ErrorMessage({ error: 'User is Banned...', description: 'The user has been banned from the tipbot and cannot use the service.\n User Banned on `' + result.banned_date + '`' });
             return;
           }
 
@@ -100,10 +126,10 @@ module.exports = {
               message.author.send({ embed })
                 .then(() => {
                   if (message.channel.type === 'dm') return;
-                  discordHelpers.errorMessage({ error: 'User Found In System...', description: 'You\'re signed up already. :thumbsup:\nTry `' + config.discord.prefix + 'help`' }, messageString);
+                  ErrorMessage({ error: 'User Found In System...', description: 'You\'re signed up already. :thumbsup:\nTry `' + config.discord.prefix + 'help`' });
                 })
                 .catch(e => {
-                  discordHelpers.errorMessage({ error: 'Direct Message Disabled', description: 'It seems you have DM\'s blocked, please enable and try again.  ' + e.message }, messageString);
+                  ErrorMessage({ error: 'Direct Message Disabled', description: 'It seems you have DM\'s blocked, please enable and try again.  ' + e.message });
                 });
             });
           });
@@ -111,7 +137,7 @@ module.exports = {
         }
         else if (found === 'false') {
           // user is not found in database. Do things here to add them
-          discordHelpers.ReplyMessage('Be right back, generating a new quantum secure address for you to send tips from...', messageString);
+          ReplyMessage('Be right back, generating a new quantum secure address for you to send tips from...');
           // Create user wallet
           const qrlWal = wallet.CreateQRLWallet;
           const WalletPromise = qrlWal();
@@ -153,7 +179,7 @@ module.exports = {
                   send_future_tip(future_tip).then(function(futureTip) {
                     const futureTipOut = JSON.parse(futureTip);
                     const tx_hash = futureTipOut.tx.transaction_hash;
-                    discordHelpers.ReplyMessage('Someone sent a tip before you signed up!\n`' + futureTipPretty + ' qrl` on the way, look for them once the transaction is confirmed by the network. `' + config.discord.prefix + 'bal` to check your wallet balance.', messageString);
+                    ReplyMessage('Someone sent a tip before you signed up!\n`' + futureTipPretty + ' qrl` on the way, look for them once the transaction is confirmed by the network. `' + config.discord.prefix + 'bal` to check your wallet balance.');
                     // write to transactions db
                     const tip_id = 1337;
                     const txInfo = { tip_id: tip_id, tx_hash: tx_hash, tx_type: 'tip' };
@@ -185,7 +211,7 @@ module.exports = {
                 message.author.send({ embed })
 
                   .catch(e => {
-                    discordHelpers.ErrrorMessage({ error: 'Direct Message Disabled', description: 'It seems you have DM\'s blocked, please enable and try again. ' + e.message }, messageString);
+                    ErrorMessage({ error: 'Direct Message Disabled', description: 'It seems you have DM\'s blocked, please enable and try again. ' + e.message });
                   }).then(() => {
                     message.author.send(`
 __**TipBot Terms and Conditions**__
@@ -204,7 +230,7 @@ __**You assume all risk by using this service**__
 
                     `)
                       .catch(e => {
-                        discordHelpers.ErrrorMessage({ error: 'Direct Message Disabled', description: 'It seems you have DM\'s blocked, please enable and try again. ' + e.message }, messageString);
+                        ErrorMessage({ error: 'Direct Message Disabled', description: 'It seems you have DM\'s blocked, please enable and try again. ' + e.message });
                       }).then(function() {
                         message.author.send(`
 :exclamation: __**RULES**__ :exclamation:
@@ -223,13 +249,13 @@ __**You assume all risk by using this service**__
                     `);
                       })
                       .catch(e => {
-                        discordHelpers.ErrrorMessage({ error: 'Direct Message Disabled', description: 'It seems you have DM\'s blocked, please enable and try again...' + e.message }, messageString);
+                        ErrorMessage({ error: 'Direct Message Disabled', description: 'It seems you have DM\'s blocked, please enable and try again...' + e.message });
                       });
-                    discordHelpers.ReplyMessage(':white_check_mark: You\'re signed up! A small drip from the faucet is on it\'s way to your new tipbot wallet as a small thanks for being a part of this community!\n Please `' + config.discord.prefix + 'agree` to my terms sent to you in DM to begin using the bot.\n*It may take a few minutes for your wallet to be created and funds deposited.*', message);
+                    ReplyMessage(':white_check_mark: You\'re signed up! A small drip from the faucet is on it\'s way to your new tipbot wallet as a small thanks for being a part of this community!\n Please `' + config.discord.prefix + 'agree` to my terms sent to you in DM to begin using the bot.\n*It may take a few minutes for your wallet to be created and funds deposited.*', message);
                   })
                   .catch(error => {
                     console.error(`Could not send help DM to ${message.author.tag}.\n`, error);
-                    discordHelpers.ErrrorMessage({ error: 'Direct Message Disabled', description: 'It seems you have DM\'s blocked, please enable and try again...' + error.message }, messageString);
+                    ErrorMessage({ error: 'Direct Message Disabled', description: 'It seems you have DM\'s blocked, please enable and try again...' + error.message });
                   });
                 message.react('🇶')
                   .then(() => message.react('🇷'))
