+++ _scripts/discord/discord-helpers.js
@@ -1,4 +1,3 @@
-const Discord = require('discord.js');
 
 
 function CheckValidChars(userName) {
@@ -12,35 +11,10 @@ function CheckValidChars(userName) {
 }
 
 
-// use to send a reply to user with delay and stop typing
-// ReplyMessage(' Check your DM\'s');
-function ReplyMessage(content, message) {
-  console.log(JSON.stringify(message));
-  message.channel.startTyping();
-  setTimeout(function() {
-    message.reply(content);
-    message.channel.stopTyping(true);
-  }, 500);
-}
 
-// errorMessage({ error: 'Can\'t access faucet from DM!', description: 'Please try again from the main chat, this function will only work there.' });
-function errorMessage(content, footer = '  .: Tipbot provided by The QRL Contributors :.', message) {
-  console.log(`discord-hellpers: ${JSON.stringify(message)}`);
-  message.channel.startTyping();
-  setTimeout(function() {
-    const embed = new Discord.MessageEmbed()
-      .setColor(0x000000)
-      .setTitle(':warning:  ERROR: ' + content.error)
-      .setDescription(content.description)
-      .setFooter(footer);
-    message.reply({ embed });
-    message.channel.stopTyping(true);
-  }, 1000);
-}
 
 
 module.exports = {
   CheckValidChars : CheckValidChars,
-  ReplyMessage : ReplyMessage,
-  errorMessage : errorMessage
+
 };
